+++ uxnanmobile/lib/presentation/theme/typography.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:uxnan/presentation/theme/colors.dart';
 
 /// Centralized typography tokens for the Uxnan design system.
 ///
@@ -20,7 +19,6 @@ class UxnanTypography {
     fontFamily: fontFamily,
     fontSize: 32,
     fontWeight: FontWeight.w700,
-    color: UxnanColors.onSurface,
     letterSpacing: -0.5,
   );
 
@@ -29,7 +27,6 @@ class UxnanTypography {
     fontFamily: fontFamily,
     fontSize: 20,
     fontWeight: FontWeight.w600,
-    color: UxnanColors.onSurface,
   );
 
   /// Small title style.
@@ -37,7 +34,6 @@ class UxnanTypography {
     fontFamily: fontFamily,
     fontSize: 14,
     fontWeight: FontWeight.w500,
-    color: UxnanColors.onSurface,
   );
 
   /// Default body style.
@@ -45,7 +41,6 @@ class UxnanTypography {
     fontFamily: fontFamily,
     fontSize: 14,
     fontWeight: FontWeight.w400,
-    color: UxnanColors.onSurface,
     height: 1.5,
   );
 
@@ -54,7 +49,6 @@ class UxnanTypography {
     fontFamily: fontFamily,
     fontSize: 12,
     fontWeight: FontWeight.w400,
-    color: UxnanColors.onSurfaceMuted,
   );
 
   /// Monospaced body style for code blocks.
@@ -62,7 +56,6 @@ class UxnanTypography {
     fontFamily: monoFontFamily,
     fontSize: 13,
     fontWeight: FontWeight.w400,
-    color: UxnanColors.onSurface,
     height: 1.6,
   );
 
@@ -71,6 +64,5 @@ class UxnanTypography {
     fontFamily: monoFontFamily,
     fontSize: 11,
     fontWeight: FontWeight.w400,
-    color: UxnanColors.onSurfaceMuted,
   );
 }